requirements.txt
python-dotenv>=1.0.0
typer>=0.9.0
comfy-cli
rich>=13.0.0
requests>=2.31.0

# Optional dependencies
# Required for workflow_to_api_json.py (workflow conversion utility):
# playwright>=1.41.0  # Also requires: playwright install 